+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "instant-mongo" extension will be documented in this 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [0.0.3] - 2020-04-07
+
+- Prints DB info in an editor window
+- Show status bar feedback while spinning up the database
+
 ## [0.0.1] - 2020-04-07
 
 - Initial release